fourth try

+++ GenThree.java
@@ -4,13 +4,13 @@
  */
 public class GenThree {
 	public static void main(String[] args) {
-        double a= Double.parseDouble(args[0]);
-        double b= Double.parseDouble(args[1]);
-        double c= Math.min (a,b);
-        double d= Math.max (a,b);
-        double x1 = ((Math.random() * (d-c)) + c);
-        double x2 = ((Math.random() * (d-c)) + c);
-        double x3 = ((Math.random() * (d-c)) + c);
+        int a= Integer.parseInt(args[0]);
+        int b= Integer.parseInt(args[1]);
+        int c= Math.min (a,b);
+        int d= Math.max (a,b);
+        int x1 = (int)((Math.random() * (d-c)) + c);
+        int x2 = (int)((Math.random() * (d-c)) + c);
+        int x3 = (int)((Math.random() * (d-c)) + c);
         System.out.println(x1);
         System.out.println(x2);
         System.out.println(x3);
